Use permissions microservice to determine user permissions
If `config.auth.permissions` is set in an app then the `req.user.can(task)` method will query the service at that endpoint with the task, and use the response status to determine the permissions for that user.

If no permissions endpoint is set then `req.user.can` will always reject. This is to ensure that permissions will not accidentally leak if the app is misconfigured.