Destroy session on /logout
We need to ensure any relics of a user's session are removed from memory when they log out, so that a new user logging in doesn't see leftover bits of their data.

Need to destroy the session on /logout, and then redirect to a special "keycloak logout" endpoint in a separate request because keycloak freaks out if it is called with no session at all.