Add mouse favicon
I was mostly just getting annoyed with the number of 404s for favicon.ico in the server logs, so I pulled up the ticket to add the mouse as a favicon. It kind of sucks, and we'll probably want to use a vector version long term, but it does the trick for now.