Use base64 encoding of INITIAL_STATE to prevent html breaking
The previous tactic of using `striptags` to remove html-like content and prevent the page source breaking when using `dangerouslySetInnerHtml` had the unfortunate side-effect of removing legitimate content from users' data. In particularly any combination of less-than and greater-than symbols in a PPL application would result in removal of all content in between.

Instead base64 encode the complete JSON and then decode on the client. This allows for legitimate html-like content in PPL applications, but removes the risk of that breaking the initial page render.