Fix `token` vs `access_token` inconsistency
The user object was being used inconsistently in different contexts, where some had the token stored in `user.token` and others in `user.access_token`. This meant that the `user.refreshProfile` function was being called without the access token available as it expected the property to be `token` where it was in fact `access_token`. This meant that all POST or PUT requests failed with a 403.

Standardise on `access_token` everywhere.