Upgrade to webpack@5
Webpack is a constant source of audit issues, none of which we should really care about as it's not a runtime dependency. However, because it was installed here as a prod dependency it was in the production dependency tree of all projects that use service (i.e. all of them).

By upgrading we get rid of a lot of the alerts from webpack@4, and by moving it to a peer dependency it can be installed where relevant as a dev dependency in parent projects and so allow it to be ignored with the --production flag on audit.

Add some fallbacks for node built-ins to the default config as these are no longer automatically included in webpack@5